Prefer class << self over module_function for function modules
- AGENTS.md: 함수 모듈 self 스타일을 class << self 로 통일 (module_function/extend self 배제)
- Move Profiles::PolymorphicActivity to app/functions/ (함수 모듈은 app/functions/ 위치 규칙)
- module_function → class << self (app/functions/ 기존 컨벤션과 일치)
- Regenerate RBS sig to mirror new path

테스트: bin/rails test test/controllers/profiles_controller_test.rb 통과 (26 runs, 0 failures)
RuboCop: 위반 없음
