+++ AGENTS.md
@@ -182,17 +182,17 @@ AI 에이전트를 위한 프로젝트 룰북입니다.
 - 변경 작업 전 `Article`, `Site`, `User`, `Comment`의 역할과 제약을 먼저 확인하고 영향 범위를 검토한다.
 - Ruby 코드에 타입 힌트를 추가하거나 수정할 때는 inline RBS 스타일을 사용한다.
 - RED, GREEN 테스트주도개발 방법에 따라 구현 코드보다 먼저 테스트 코드를 작성한 후 구현 코드를 작성한다.
-- 서비스 객체를 만들거나 수정할 때는 기존 `OperationService` 을 상속하여 ROP 패턴을 따른다. 단순한 유틸리티성 서비스는 module function으로 작성해도 된다.
+- 서비스 객체를 만들거나 수정할 때는 기존 `OperationService` 을 상속하여 ROP 패턴을 따른다. 단순한 유틸리티성 서비스는 `class << self` 를 사용한 함수 모듈로 작성한다.
 - 소셜 미디어 연동 코드는 `SocialMediaService` 기반 구조와 플랫폼별 서비스 분리를 유지한다.
 - 변경을 마무리하기 전에 테스트 여부와 미실행 사유를 명확히 남긴다.
 - 관련 배경 문서가 필요하면 `docs/CLAUDE_WORKFLOW.md`, `docs/postgresql-extensions.md`를 우선 참고한다.
 - 뷰 클래스는 `Views::Base`를, 컴포넌트 클래스는 `Components::Base`를 상속한다.
 - `OperationService`(`Dry::Operation`)의 `call` 메서드에서 `return Failure(:x)`를 직접 반환하면 `Dry::Operation`이 이를 `Success(Failure(:x))`로 감싸버린다. `Failure`를 반환하려면 반드시 `step`을 통해야 한다. guard clause도 `step validate_something(...)` 형태로 호출한다.
 - **`class` 대신 `module` 우선 원칙** (ref: Dave Thomas, "Eliminating the `class` Keyword from Ruby"):
-  - **오브젝트 팩토리가 아니면 module이다.** 인스턴스 변수, 인스턴스 메서드, 생성자가 없다면 `class`가 아니라 `module`(+ `module_function`)로 작성한다.
+  - **오브젝트 팩토리가 아니면 module이다.** 인스턴스 변수, 인스턴스 메서드, 생성자가 없다면 `class`가 아니라 `module`(+ `class << self`)로 작성한다. 함수 모듈의 self 스타일은 `class << self` 로 통일한다(`module_function`·`extend self`는 쓰지 않는다).
   - **추상 기반 클래스(abstract base class)는 Ruby에 필요 없다.** 상속으로 메서드를 주입하는 대신 mixin(`include`/`extend`/`prepend`)을 사용한다. (Rails 프레임워크의 `ApplicationRecord`, `ApplicationController` 등은 예외)
   - **`new` 직후 invalid한 상태면 안 된다.** 생성자에서 모든 필수 값을 받아 유효한 객체만 만들어야 한다. `new` 후 setter 호출이 필요하다면 클래스 설계가 잘못된 것이다.
-  - **`module_function` 유틸리티의 위치는 `app/functions/`다.** `app/services/`는 상태를 가진 서비스 객체, `OperationService` 상속 객체, 또는 인스턴스 기반 협력 객체에만 사용한다. 단순 함수성 모듈을 습관적으로 `services` 아래 두지 않는다.
+  - **함수 모듈(`class << self`)의 위치는 `app/functions/`다.** `app/services/`는 상태를 가진 서비스 객체, `OperationService` 상속 객체, 또는 인스턴스 기반 협력 객체에만 사용한다. 단순 함수성 모듈을 습관적으로 `services`·`models` 아래 두지 않는다.
   - **함수 모듈 이름에 불필요한 `_service` postfix를 붙이지 않는다.** 함수 모듈은 `RegistrationService`, `CallbackService`보다 `Registration`, `Callbacks`, `UserMatcher`처럼 역할이 바로 드러나는 이름을 사용한다.
   - **함수 모듈의 진입점 이름을 무조건 `call`로 짓지 않는다.** `call`은 서비스 객체 관용구에 가깝다. `app/functions/` 아래의 함수 모듈은 `match_user`, `suggest_username`, `build_auth_result`, `register_user`처럼 도메인 의미가 드러나는 메서드명을 우선 사용한다.
 - **여러 값을 묶어 반환·전달할 때는 untyped Hash 대신 불변 값 객체(`Data.define`)를 쓴다.**
